requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.3
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==25.2
drf-spectacular==0.28.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0

reportlab
hypothesis

# Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
pytest-xdist==3.6.1
